refactor: 리듀서에서 products, order 정보 삭제
