+++ src/pages/Order/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { useHistory } from 'react-router';
 import { useSelector } from 'react-redux';
+
 import API from '../../request/api';
 import { Button, HighlightText, Product } from '../../components/shared';
 import { COLOR, MESSAGE, PATH } from '../../constants';
