+++ src/pages/ProductDetail/index.js
@@ -1,19 +1,19 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
 import { useHistory, useParams } from 'react-router';
-import { Button } from '../../components/shared';
-import Thumbnail from '../../components/shared/Thumbnail';
-import { COLOR, FETCH_URL, MESSAGE, PATH } from '../../constants';
-import API from '../../request/api';
-import useFetch from '../../request/useFetch';
-import { addItemToCart } from '../../store/cartReducer';
-import { Container, ProductSummary, Price } from './style';
 
-import noProduct from '../../assets/images/no_product.png';
+import API from '../../request/api';
+import useFetch from '../../hooks/useFetch';
+import { addItemToCart } from '../../store/cartReducer';
+
+import { COLOR, FETCH_URL, MESSAGE, PATH } from '../../constants';
+import { Button, Thumbnail } from '../../components/shared';
+import { Container, ProductSummary, Price } from './style';
+import noProductImage from '../../assets/images/no_product.png';
 
 const ProductDetail = () => {
   const { id } = useParams();
-  const [item, itemError] = useFetch(FETCH_URL.GET_PRODUCT_DETAIL(id));
+  const [item, itemFetchError] = useFetch(FETCH_URL.GET_PRODUCT_DETAIL(id));
 
   const history = useHistory();
   const dispatch = useDispatch();
@@ -22,10 +22,10 @@ const ProductDetail = () => {
     history.push(PATH.MAIN);
   };
 
-  if (itemError) {
+  if (itemFetchError) {
     return (
       <Container>
-        <img src={noProduct} alt="상품 상세 없음" />
+        <img src={noProductImage} alt="상품 상세 없음" />
         <p>존재하지 않는 상품입니다</p>
         <Button type="button" size="medium" backgroundColor={COLOR['GRAY-300']} onClick={goMain}>
           메인으로 돌아가기
